Update to functional TBasket code
Can properly load a TTree and deserialize/decompress individual
TBaskets. Want to do some refactoring around exception handling before
continuing further, so this is a good point to checkpoint before I start
ripping out throws exceptions.

+++ src/main/java/edu/vanderbilt/accre/laurelin/root_proxy/FileInterface.java
@@ -5,7 +5,7 @@ import java.nio.ByteBuffer;
 import java.util.concurrent.Future;
 
 public interface FileInterface {
-	public ByteBuffer read(long offset, int len) throws IOException;
+	public ByteBuffer read(long offset, long len) throws IOException;
 	public ByteBuffer[] readv(int[] offsets, int[] lens) throws IOException;
 	public Future<ByteBuffer> readAsync(int offset, int len) throws IOException;
 	public Future<ByteBuffer>[] readvAsync(int[] offsets, int[] lens) throws IOException;
